+++ bugsnag-android-performance/src/test/java/com/bugsnag/android/performance/internal/BugsnagClockTest.kt
@@ -1,10 +1,12 @@
 package com.bugsnag.android.performance.internal
 
 import android.os.SystemClock
+import com.bugsnag.android.performance.test.withStaticMock
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.kotlin.doReturn
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
@@ -18,5 +20,11 @@ class BugsnagClockTest {
     fun testCurrentTime() {
         val elapsedRealtimeNanos = SystemClock.elapsedRealtimeNanos()
         assertTrue(BugsnagClock.elapsedNanosToUnixTime(elapsedRealtimeNanos) > elapsedRealtimeNanos)
+    }
+
+    @Test
+    fun testSafeBootFallback_Zero() = withStaticMock<SystemClock> { mockedClock ->
+        mockedClock.`when`<Long> { SystemClock.elapsedRealtimeNanos() } doReturn Long.MAX_VALUE
+        assertEquals(0L, BugsnagClock.safeBootTimeNanos())
     }
 }