+++ features/fixtures/mazeracer/app/src/main/java/com/bugsnag/mazeracer/scenarios/ActivityLoadInstrumentationScenario.kt
@@ -3,16 +3,14 @@ package com.bugsnag.mazeracer.scenarios
 import com.bugsnag.android.performance.AutoInstrument
 import com.bugsnag.android.performance.BugsnagPerformance
 import com.bugsnag.android.performance.PerformanceConfiguration
-import com.bugsnag.android.performance.internal.InternalDebug
 import com.bugsnag.mazeracer.ActivityViewLoadActivity
 import com.bugsnag.mazeracer.Scenario
 
 class ActivityLoadInstrumentationScenario(
     config: PerformanceConfiguration,
-    scenarioMetadata: String
+    scenarioMetadata: String,
 ) : Scenario(config, scenarioMetadata) {
     init {
-        InternalDebug.spanBatchSizeSendTriggerPoint = 2
         config.autoInstrumentActivities =
             scenarioMetadata.takeIf { it.isNotBlank() }
             ?.let { AutoInstrument.valueOf(it) }
@@ -21,8 +19,11 @@ class ActivityLoadInstrumentationScenario(
 
     override fun startScenario() {
         BugsnagPerformance.start(config)
-        context.startActivity(
-            ActivityViewLoadActivity.intent(context, config.autoInstrumentActivities)
+        startActivityAndFinish(
+            ActivityViewLoadActivity.intent(
+                context,
+                config.autoInstrumentActivities,
+            ),
         )
     }
 }